requirements.txt
langchain
langchain-community
chromadb
pypdf
python-dotenv
ipykernel
sentence-transformers
langchain-chroma
langchain-openai